+++ addons/CSScriptAddon/CSScriptAddon.vb
@@ -13,8 +13,8 @@ Public Class CSScriptAddon
     Sub New()
         Dim scriptFiles As New List(Of String)
 
-        If Directory.Exists(mp.MpvConfFolder + "scripts") Then
-            scriptFiles.AddRange(Directory.GetFiles(mp.MpvConfFolder + "scripts", "*.cs"))
+        If Directory.Exists(mp.ConfFolder + "scripts") Then
+            scriptFiles.AddRange(Directory.GetFiles(mp.ConfFolder + "scripts", "*.cs"))
         End If
 
         If Directory.Exists(Application.StartupPath + "\scripts") Then