+++ Changelog.md
@@ -1,5 +1,11 @@
 ###
 
+###
+
+- the help in the config editor was improved
+- clipboard monitoring for URLs can be disabled in the settings
+- internal refactoring that might break features, scripts or extensions
+
 ### 4.1
 
 - drag & drop support for subtitles was added